requirements.txt
pyqrcode==1.2
websockets==10.0
bson==0.5.10
pygame==2.1.0
pyopengl==3.1.5
pyopengl-accelerate==3.1.5
pywavefront==1.3.3
